Updated Blue Robot Code
I corrected the outdated coordinates. They may be slightly wrong but they are close enough to what I remember them being at the end of the lecture. The main issue was not the robot head being the wrong length, it was that the robot was not centred properly or calibrated.